feat: 保存进度

+++ src/SuperForm/ArrayWrapper.tsx
@@ -1,14 +1,23 @@
+/**
+ * ArrayWrapper对于Array类型的组件封装一层
+ * 提供addItem、deleteItem和items，使得子组件使用更简单
+ */
 import React, { useState, useCallback } from 'react';
 import { GetArrayItem, ComponentType } from './type';
 
 interface IArrayWrapperProps {
   getArrayItem: GetArrayItem; // 新增子项
   initialValue?: Array<any>;
-  Widget:  ComponentType;
+  WidgetClass?:  ComponentType;
 }
 
 export default React.memo((props: IArrayWrapperProps) => {
-  const { initialValue = [], getArrayItem, Widget, ...other } = props;
+  const {
+    initialValue = [],
+    getArrayItem,
+    WidgetClass,
+    ...other
+  } = props;
   const [items, setItems] = useState<Array<number>>(initialValue.map((item, index) => index));
   
   const addItem = useCallback(() => {
@@ -20,17 +29,20 @@ export default React.memo((props: IArrayWrapperProps) => {
     items.splice(index, 1);
     setItems([...items]);
   }, [items]);
-  
-  return Widget && (
-    <Widget
-      {...{
-        ...other,
-        addItem,
-        deleteItem,
-        items: items.map((index) => {
-          return getArrayItem(index);
-        }),
-      }}
-    />
-  );
+ 
+  if (WidgetClass) {
+    return (
+      <WidgetClass
+        {...{
+          ...other,
+          addItem,
+          deleteItem,
+          items: items.map((index) => {
+            return getArrayItem(index);
+          }),
+        }}
+      />
+    )
+  }
+  return null;
 });
